Perform CAESAR embedding of Cells Using FAST with Spatial Weights

This function computes cell embedding using the CAESAR framework with FAST for dimensionality reduction and spatial adjacency weights. It integrates variable feature selection and spatial adjacency information to generate low-dimensional representations for cells.

cellembedding_seurat(
  seu,
  adjm,
  assay = NULL,
  slot = "data",
  nfeatures = 2000,
  q = 50,
  reduction.name = "caesar",
  var.features = NULL,
  ...
)

Arguments

seu

A Seurat object. The Seurat object should contain gene expression data and be preprocessed with variable features identified.

adjm

A spatial adjacency matrix representing the relationships between cells or spots in spatial transcriptomic data.

assay

A character string specifying which assay to use from the Seurat object. If NULL, the function will use the default assay.

slot

The data slot to use for feature extraction (e.g., "data", "counts"). Default is "data".

nfeatures

The number of features to select for analysis. Default is 2000.

q

An integer specifying the number of dimensions for the reduced embeddings. Default is 50.

reduction.name

A character string specifying the name for the dimensional reduction. Default is "caesar".

var.features

A vector of variable features (genes) to use for the embedding. If NULL, the function will use variable features stored in the Seurat object.

...

Additional arguments passed to `FAST_run`.

Value

The modified Seurat object with the co-embedding results (cell and gene embeddings) stored in the specified dimensional reduction slot.

See also

Examples

data(toydata)

adjm <- ProFAST::AddAdj(as.matrix(pos))
#> The spatial cooridnates are 2 dimensions
#> Find the adjacency matrix by bisection method...
#> Current radius is 43.54
#> Median of neighborhoods is 0
#> Current radius is 64.81
#> Median of neighborhoods is 3
#> Current radius is 64.81
#> Median of neighborhoods is 7
seu <- cellembedding_seurat(
    seu = seu,
    adjm = adjm
)
#> Finish variable intialization 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 2, loglik= -593056.539819, dloglik=0.999724 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 3, loglik= -570874.007889, dloglik=0.037404 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 4, loglik= -566704.048192, dloglik=0.007305 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 5, loglik= -564254.720016, dloglik=0.004322 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 6, loglik= -562690.648571, dloglik=0.002772 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 7, loglik= -561509.654897, dloglik=0.002099 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 8, loglik= -560641.895001, dloglik=0.001545 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 9, loglik= -559975.664008, dloglik=0.001188 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 10, loglik= -559444.033359, dloglik=0.000949 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 11, loglik= -558992.103674, dloglik=0.000808 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 12, loglik= -558589.244723, dloglik=0.000721 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 13, loglik= -558214.060529, dloglik=0.000672 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 14, loglik= -557856.061807, dloglik=0.000641 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 15, loglik= -557510.040282, dloglik=0.000620 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 16, loglik= -557174.586610, dloglik=0.000602 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 17, loglik= -556849.982492, dloglik=0.000583 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 18, loglik= -556538.525733, dloglik=0.000559 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 19, loglik= -556243.551873, dloglik=0.000530 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 20, loglik= -555968.285798, dloglik=0.000495 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 21, loglik= -555714.627018, dloglik=0.000456 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 22, loglik= -555483.143937, dloglik=0.000417 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 23, loglik= -555273.458573, dloglik=0.000377 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 24, loglik= -555084.680514, dloglik=0.000340 
#> Satrt ICM and E-step! 
#> Finish ICM and E-step! 
#> iter = 25, loglik= -554915.482787, dloglik=0.000305 
#> 2024-09-02 13:45:48.826839 : ***** Finish calculate CAESAR embedding, 0.074 mins elapsed.
print(seu)
#> An object of class Seurat 
#> 302 features across 3000 samples within 1 assay 
#> Active assay: RNA (302 features, 302 variable features)
#>  2 dimensional reductions calculated: caesar, pos